# openapi.yaml
openapi: 3.0.1
info:
  title: Terraware Seed Bank
  description: Local server API for seed banks
  license:
    name: MIT
  version: 0.1-SNAPSHOT
servers:
- url: http://localhost:32109
  description: Generated server url
...
paths:
...
  /api/v1/seedbank/values/species:
    get:
      tags:
      - SeedBankApp
      description: Use /api/v1/species instead.
      operationId: listSpecies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSpeciesResponsePayload'
      deprecated: true
    post:
      tags:
      - SeedBankApp
      description: Use /api/v1/species instead.
      operationId: createSpecies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSpeciesResponsePayload'
        "409":
          description: A species with the requested name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSpeciesResponsePayload'
      deprecated: true
  /api/v1/seedbank/values/species/{id}:
    post:
      tags:
      - SeedBankApp
      summary: Updates an existing species.
      description: "If the species is being renamed and the species name in the payload\
        \ already exists, the existing species replaces the one in the request (thus\
        \ merging the requested species into the one that already had the name) and\
        \ its ID is returned."
      operationId: updateSpecies
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species updated or merged with an existing species.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateSpeciesResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
...
  /api/v1/species:
    get:
      tags:
      - GISApp
      summary: Lists all known species.
      operationId: speciesList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesListResponsePayload'
    post:
      tags:
      - GISApp
      summary: Creates a new species.
      operationId: speciesCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesCreateResponsePayload'
        "409":
          description: A species with the requested name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesCreateResponsePayload'
  /api/v1/species/names:
    get:
      tags:
      - GISApp
      summary: Lists all species names.
      operationId: speciesNamesListAll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesNamesListResponsePayload'
    post:
      tags:
      - GISApp
      summary: Adds a new name for an existing species.
      operationId: speciesNameCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesNameRequestPayload'
        required: true
      responses:
        "200":
          description: Species name added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesNameCreateResponsePayload'
        "404":
          description: The species does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
        "409":
          description: The species already has the requested name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesNameCreateResponsePayload'
  /api/v1/species/names/{speciesNameId}:
    get:
      tags:
      - GISApp
      description: Gets information about a single species name.
      operationId: speciesNameGet
      parameters:
      - name: speciesNameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species name retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesNameGetResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    put:
      tags:
      - GISApp
      description: Updates one of the names of a species.
      operationId: speciesNameUpdate
      parameters:
      - name: speciesNameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesNameRequestPayload'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
    delete:
      tags:
      - GISApp
      description: Deletes one of the secondary names of a species.
      operationId: speciesNameDelete
      parameters:
      - name: speciesNameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species name deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
        "409":
          description: Cannot delete the primary name of a species.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
  /api/v1/species/{speciesId}:
    get:
      tags:
      - GISApp
      summary: Gets information about a single species.
      operationId: speciesRead
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesGetResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    put:
      tags:
      - GISApp
      summary: Updates an existing species.
      operationId: speciesUpdate
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesRequestPayload'
        required: true
      responses:
        "200":
          description: Species updated or merged with an existing species.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
    delete:
      tags:
      - GISApp
      summary: Delete an existing species.
      operationId: speciesDelete
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
        "409":
          description: Cannot delete the species because it is currently in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponsePayload'
  /api/v1/species/{speciesId}/names:
    get:
      tags:
      - GISApp
      operationId: speciesNamesList
      parameters:
      - name: speciesId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Species names retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesNamesListResponsePayload'
        "404":
          description: The species does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorResponsePayload'
components:
  schemas:
...
    CreateSpeciesRequestPayload:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    CreateSpeciesResponsePayload:
      required:
      - details
      - status
      type: object
      properties:
        details:
          $ref: '#/components/schemas/SpeciesDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    ErrorDetails:
      required:
      - message
      type: object
      properties:
        message:
          type: string
...
    ListSpeciesResponsePayload:
      required:
      - status
      - values
      type: object
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/SpeciesDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
...
    SimpleErrorResponsePayload:
      required:
      - error
      - status
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetails'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SimpleSuccessResponsePayload:
      required:
      - status
      type: object
      properties:
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesCreateResponsePayload:
      required:
      - id
      - status
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesDetails:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SpeciesGetResponsePayload:
      required:
      - species
      - status
      type: object
      properties:
        species:
          $ref: '#/components/schemas/SpeciesResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesListResponsePayload:
      required:
      - species
      - status
      type: object
      properties:
        species:
          type: array
          items:
            $ref: '#/components/schemas/SpeciesResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesNameCreateResponsePayload:
      required:
      - id
      - status
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesNameGetResponsePayload:
      required:
      - speciesName
      - status
      type: object
      properties:
        speciesName:
          $ref: '#/components/schemas/SpeciesNamesResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesNameRequestPayload:
      required:
      - name
      - speciesId
      type: object
      properties:
        isScientific:
          type: boolean
          description: True if name is a scientific name for the species.
        locale:
          type: string
        name:
          type: string
        speciesId:
          type: integer
          format: int64
    SpeciesNamesListResponsePayload:
      required:
      - speciesNames
      - status
      type: object
      properties:
        speciesNames:
          type: array
          items:
            $ref: '#/components/schemas/SpeciesNamesResponseElement'
        status:
          $ref: '#/components/schemas/SuccessOrError'
    SpeciesNamesResponseElement:
      required:
      - id
      - isScientific
      - name
      - speciesId
      type: object
      properties:
        id:
          type: integer
          format: int64
        isScientific:
          type: boolean
          description: True if name is the scientific name for the species.
        name:
          type: string
        speciesId:
          type: integer
          format: int64
    SpeciesRequestPayload:
      required:
      - name
      type: object
      properties:
        conservationStatus:
          type: string
        familyId:
          type: integer
          format: int64
        isScientific:
          type: boolean
          description: True if name is the scientific name for the species.
        name:
          type: string
        plantForm:
          type: string
          enum:
          - Tree
          - Shrub
          - Vine
          - Liana
          - Herbaceous
        rare:
          type: string
          enum:
          - "No"
          - "Yes"
          - Unsure
        tsn:
          type: string
          description: Taxonomic serial number from ITIS database.
          externalDocs:
            url: https://en.wikipedia.org/wiki/Integrated_Taxonomic_Information_System
    SpeciesResponseElement:
      required:
      - id
      - isScientific
      - name
      type: object
      properties:
        conservationStatus:
          type: string
        familyId:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        isScientific:
          type: boolean
          description: True if name is the scientific name for the species.
        name:
          type: string
        plantForm:
          type: string
          enum:
          - Tree
          - Shrub
          - Vine
          - Liana
          - Herbaceous
        rare:
          type: string
          enum:
          - "No"
          - "Yes"
          - Unsure
        tsn:
          type: string
          description: Taxonomic serial number from ITIS database.
          externalDocs:
            url: https://en.wikipedia.org/wiki/Integrated_Taxonomic_Information_System
...
    SuccessOrError:
      type: string
      description: Indicates of success or failure of the requested operation.
      enum:
      - ok
      - error
...
    UpdateSpeciesResponsePayload:
      required:
      - status
      type: object
      properties:
        mergedWithSpeciesId:
          type: integer
          description: "If the requested species name already existed, the ID of the\
            \ existing species. Will not be present if the requested species name\
            \ did not already exist."
          format: int64
        status:
          $ref: '#/components/schemas/SuccessOrError'
...